refactor(v0): extract session state read model
